--- src/App.css
@@ -1,7 +1,0 @@
-/* #root {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-} */
-
