+++ src/components/Logo.jsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+export const Logo = () => {
+  return (
+    <>
+      <a
+        href="https://www.frro.utn.edu.ar/"
+        target="_blank"
+        rel="noopener noreferrer"
+      >
+        <img
+          src="./src/assets/descarga.png"
+          alt=""
+          className="absolute lg:max-w-28 lg:block md:block md:max-w-20 max-w-16 top-3 left-3 lg:top-10 lg:left-10"
+        />
+      </a>
+    </>
+  );
+};
